server/service/requirements.txt
pycrypto==2.6.1
flask-login==0.4.1
aip==1.0.0.5
amqp==2.3.2
avro==1.8.2
backports.functools-lru-cache==1.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
billiard==3.5.0.4
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cos-python-sdk-v5==1.4.0
cycler==0.10.0
dicttoxml==1.7.4
docker==2.6.1
docker-pycreds==0.2.1
docopt==0.6.2
easydict==1.7
elasticsearch==6.2.0
enum34==1.1.6
Flask==0.12
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
futures==3.2.0
glob2==0.6
grpcio==1.8.1
grpcio-tools==1.8.1
icrawler==0.4.9
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.1
mysqlclient==1.3.12
nose==1.3.7
numpy==1.13.3
olefile==0.44
opencv-python==3.4.0.12
pandas==0.21.1
Pillow==5.1.0
pipreqs==0.4.9
pprint==0.1
protobuf==3.5.0.post1
pymongo==3.5.1
PyMySQL==0.8.0
pyparsing==2.2.0
pytesseract==0.2.0
python-dateutil==2.6.1
pytz==2017.3
qcloud-cos-v4==0.0.22
requests==2.18.4
selenium==3.7.0
six==1.11.0
subprocess32==3.2.7
urllib3==1.22
vin-decoder==0.1.0
vine==1.1.4
websocket-client==0.44.0
Werkzeug==0.11.15
yarg==0.1.9
zeroc-ice==3.7.0.1
